+++ code/math/miller-rabin.cpp
@@ -1,4 +1,4 @@
-bool is_prime(int n) {
+bool is_prime(ll n) {
     if(n < 2 || n % 6 % 4 != 1) return (n | 1) == 3;
     int A[] = {2, 325, 9375, 28178, 450775, 9780504, 1795265022},
               s = __builtin_ctzll(n - 1), d = n >> s;
